Organized files and updated imports
