+++ lib/addons/calico/index.ts
@@ -1,4 +1,4 @@
-import { ClusterAddOn, ClusterInfo } from "../../eksBlueprintStack";
+import { ClusterAddOn, ClusterInfo } from "../../stacks/eks-blueprint-stack";
 
 export class CalicoAddon implements ClusterAddOn {
