+++ lib/cluster-providers/ec2-cluster-provider.ts
