+++ lib/index.ts
@@ -1,3 +1,3 @@
-export * from './eksBlueprintStack';
+export * from './stacks/eks-blueprint-stack';
 export * as Addons from './addons';
